deleted more files and cleaned some routes up
