+++ routes/passport-api-routes.js
@@ -3,16 +3,11 @@ var db = require("../models");
 var passport = require("../config/passport");
 
 module.exports = function (app) {
-  // Using the passport.authenticate middleware with our local strategy.
-  // If the user has valid login credentials, send them to the members page.
-  // Otherwise the user will be sent an error
+
   app.post("/api/login", passport.authenticate("local"), function (req, res) {
     res.json(req.user);
   });
 
-  // Route for signing up a user. The user's password is automatically hashed and stored securely thanks to
-  // how we configured our Sequelize User Model. If the user is created successfully, proceed to log the user in,
-  // otherwise send back an error
   app.post("/api/signup", function (req, res) {
     db.User.create({
       email: req.body.email,
@@ -48,78 +43,73 @@ module.exports = function (app) {
     }
   });
   var db = require("../models");
+  app.get("/api/posts", function (req, res) {
+    var query = {};
+    if (req.query.author_id) {
+      query.UserId = req.query.user_id;
+    }
+    // 1. Add a join here to include all of the Authors to these posts
+    db.Post.findAll({
+      where: query,
+      include: [db.User]
+    }).then(function (dbPost) {
+      res.json(dbPost);
+    });
+  });
 
-  // Routes
-  // =============================================================
-  module.exports = function (app) {
+  // Get route for retrieving a single post
+  app.get("/api/posts/:id", function (req, res) {
 
-    // GET route for getting all of the posts
-    app.get("/api/posts", function (req, res) {
-      var query = {};
-      if (req.query.author_id) {
-        query.UserId = req.query.user_id;
+    db.Post.findOne({
+      where: {
+        id: req.params.id
+      },
+      include: [db.User]
+    }).then(function (dbPost) {
+      console.log(dbPost);
+      res.json(dbPost);
+    });
+  });
+
+  // POST route for saving a new post
+  app.post("/api/posts", function (req, res) {
+    db.Post.create(req.body).then(function (dbPost) {
+      res.json(dbPost);
+    });
+  });
+
+  // DELETE route for deleting posts
+  app.delete("/api/posts/:id", function (req, res) {
+    db.Post.destroy({
+      where: {
+        id: req.params.id
       }
-      // 1. Add a join here to include all of the Authors to these posts
-      db.Post.findAll({
-        where: query,
-        include: [db.User]
-      }).then(function (dbPost) {
-        res.json(dbPost);
-      });
+    }).then(function (dbPost) {
+      res.json(dbPost);
     });
+  });
 
-    // Get route for retrieving a single post
-    app.get("/api/posts/:id", function (req, res) {
-
-      db.Post.findOne({
+  // PUT route for updating posts
+  app.put("/api/posts", function (req, res) {
+    db.Post.update(
+      req.body,
+      {
         where: {
-          id: req.params.id
-        },
-        include: [db.User]
-      }).then(function (dbPost) {
-        console.log(dbPost);
-        res.json(dbPost);
-      });
-    });
-
-    // POST route for saving a new post
-    app.post("/api/posts", function (req, res) {
-      db.Post.create(req.body).then(function (dbPost) {
-        res.json(dbPost);
-      });
-    });
-
-    // DELETE route for deleting posts
-    app.delete("/api/posts/:id", function (req, res) {
-      db.Post.destroy({
-        where: {
-          id: req.params.id
+          id: req.body.id
         }
       }).then(function (dbPost) {
         res.json(dbPost);
       });
-    });
+  });
 
-    // PUT route for updating posts
-    app.put("/api/posts", function (req, res) {
-      db.Post.update(
-        req.body,
-        {
-          where: {
-            id: req.body.id
-          }
-        }).then(function (dbPost) {
-          res.json(dbPost);
-        });
+  //test route 
+  app.post("/api/signup", function (req, res) {
+    console.log(req.body.test)
+    res.json({ message: "server respons" })
+    db.User.create(req.body).then(function (dbPost) {
+      res.json(dbPost);
     });
+  })
 
-    //test route 
-    app.post("/api/signup", function (req, res) {
-      console.log(req.body.test)
-      res.json({ message: "server respons" })
-      db.User.create(req.body).then(function (dbPost) {
-        res.json(dbPost);
-      });
-    })
-  }
+  
 }
