+++ views/login.handlebars
@@ -1,5 +1,5 @@
 <h1>login</h1>
-<p>{{message}}</p>
+<p>{{login}}</p>
 <!DOCTYPE html>
 <html lang="en">
